back/requirements.txt
aiohttp==1.3.5
amqp==2.5.1
async-timeout==3.0.1
Babel==2.7.0
billiard==3.6.1.0
cairocffi==0.9.0
CairoSVG==2.1.3
celery==4.3.0
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
cx-Oracle==7.2.3
defusedxml==0.5.0
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.1
django-auth-ldap==2.0.0
django-celery-beat==1.4.0
django-celery-results==1.1.2
django-cors-headers==2.4.0
django-debug-toolbar==1.9.1
django-discover-runner==1.0
django-import-export==1.0.1
django-puppeteer-pdf==0.1.3
django-redis==4.10.0
django-rest-auth==0.9.3
django-role-permissions==2.2.0
django-stubs==1.6.0
django-timezone-field==3.1
django-wkhtmltopdf==3.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.2.0
dnspython==1.16.0
dry-rest-permissions==0.1.10
enum34==1.1.6
et-xmlfile==1.0.1
eventlet==0.25.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.11.3
idna==2.8
importlib-metadata==0.23
jdcal==1.4
kombu==4.6.5
monotonic==1.5
more-itertools==7.2.0
multidict==4.5.2
mypy==0.782
mypy-extensions==0.4.3
odfpy==1.3.6
openpyxl==2.5.5
pdfrw==0.4
Pillow==5.2.0
psycopg2-binary==2.7.5
ptvsd==4.3.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.18
pycpfcnpj==1.5.1
PyJWT==1.6.4
PyPDF2==1.26.0
Pyphen==0.9.4
python-crontab==2.3.9
python-dateutil==2.8.0
python-decouple==3.1
python-ldap==3.2.0
pytz==2018.5
PyYAML==3.13
redis==3.3.8
reportbro-fpdf==1.7.9
reportbro-lib==1.3.1
reportlab==3.5.4
requests==2.21.0
simpleeval==0.9.8
six==1.11.0
sqlparse==0.2.4
striprtf==0.0.5
tablib==0.12.1
tinycss2==0.6.1
typed-ast==1.4.0
typing==3.7.4.1
typing-extensions==3.7.4
unicodecsv==0.14.1
urllib3==1.24.1
vine==1.3.0
WeasyPrint==45
webencodings==0.5.1
whitenoise==3.3.1
xhtml2pdf==0.2.3
xlrd==1.1.0
XlsxWriter==1.2.1
xlwt==1.3.0
yarl==0.9.8
zipp==0.6.0
